Created Mini Stock Web Application
This is a mini web application created using the Python Dash library. The app fetches the stock data for the current year from Yahoo Finance using the pandas API. The user can enter the ticker symbol of any stock, which will update the candlestick chart. The user is also able to increase or decrease the date range using a slider.